[#2496] Add datasource metric collection
1. Moving class's modules
2. Moving class's package
