+++ bootstrap-core/src/main/java/com/navercorp/pinpoint/bootstrap/context/PluginMonitorRegistry.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.navercorp.pinpoint.bootstrap.plugin.monitor;
+package com.navercorp.pinpoint.bootstrap.context;
 
 /**
  * @author Taejin Koo
