+++ bootstrap-core/src/main/java/com/navercorp/pinpoint/bootstrap/context/TraceContext.java
@@ -17,7 +17,6 @@
 package com.navercorp.pinpoint.bootstrap.context;
 
 import com.navercorp.pinpoint.bootstrap.config.ProfilerConfig;
-import com.navercorp.pinpoint.bootstrap.plugin.monitor.PluginMonitorContext;
 
 /**
  * @author emeroad
