+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/context/monitor/DataSourceMonitorList.java
@@ -14,8 +14,10 @@
  * limitations under the License.
  */
 
-package com.navercorp.pinpoint.bootstrap.plugin.monitor;
+package com.navercorp.pinpoint.profiler.context.monitor;
 
+import com.navercorp.pinpoint.bootstrap.context.DataSourceMonitor;
+import com.navercorp.pinpoint.bootstrap.context.PluginMonitorRegistry;
 import com.navercorp.pinpoint.bootstrap.logging.PLogger;
 import com.navercorp.pinpoint.bootstrap.logging.PLoggerFactory;
