+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/context/monitor/DataSourceMonitorWrapperFactory.java
@@ -14,13 +14,25 @@
  * limitations under the License.
  */
 
-package com.navercorp.pinpoint.bootstrap.plugin.monitor;
+package com.navercorp.pinpoint.profiler.context.monitor;
+
+import com.navercorp.pinpoint.bootstrap.context.DataSourceMonitor;
+
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * @author Taejin Koo
  */
-public interface PluginMonitorContext {
+public class DataSourceMonitorWrapperFactory {
 
-    PluginMonitorRegistry<DataSourceMonitor> getDataSourceMonitorRegistry();
+    private final AtomicInteger idGenerator = new AtomicInteger();
+
+    public DataSourceMonitorWrapper create(DataSourceMonitor pluginMonitor) {
+        return new DataSourceMonitorWrapper(idGenerator.incrementAndGet(), pluginMonitor);
+    }
+
+    public int latestIssuedId() {
+        return idGenerator.get();
+    }
 
 }
