+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/context/monitor/PluginMonitorWrapperLocator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.navercorp.pinpoint.bootstrap.plugin.monitor;
+package com.navercorp.pinpoint.profiler.context.monitor;
 
 import java.util.List;
