+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/monitor/codahale/MetricMonitorRegistry.java
@@ -24,10 +24,10 @@ import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.jvm.GarbageCollectorMetricSet;
 import com.codahale.metrics.jvm.MemoryUsageGaugeSet;
 import com.codahale.metrics.jvm.ThreadStatesGaugeSet;
-import com.navercorp.pinpoint.bootstrap.plugin.monitor.DataSourceMonitorWrapper;
-import com.navercorp.pinpoint.bootstrap.plugin.monitor.PluginMonitorWrapperLocator;
 import com.navercorp.pinpoint.profiler.context.TransactionCounter;
 import com.navercorp.pinpoint.profiler.context.active.ActiveTraceLocator;
+import com.navercorp.pinpoint.profiler.context.monitor.DataSourceMonitorWrapper;
+import com.navercorp.pinpoint.profiler.context.monitor.PluginMonitorWrapperLocator;
 import com.navercorp.pinpoint.profiler.monitor.CounterMonitor;
 import com.navercorp.pinpoint.profiler.monitor.EventRateMonitor;
 import com.navercorp.pinpoint.profiler.monitor.HistogramMonitor;
